Fix breaking change introduced in https://github.com/CanCanCommunity/cancancan/pull/204
