+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+* [#571](https://github.com/CanCanCommunity/cancancan/pull/571): allows to check ability even the object implements `#to_a`. ([@mtsmfm][])
 * [#612](https://github.com/CanCanCommunity/cancancan/pull/612): Suppress keyword arguments warning for Ruby 2.7.0. ([@koic][])
 
 ## 3.0.2
@@ -654,4 +655,5 @@ Please read the [guide on migrating from CanCanCan 2.x to 3.0](https://github.co
 [@kaspernj]: https://github.com/kaspernj
 [@frostblooded]: https://github.com/frostblooded
 [@eloyesp]: https://github.com/eloyesp
+[@mtsmfm]: https://github.com/mtsmfm
 [@koic]: https://github.com/koic
